Refactor sensor runner in st2reactor   * Add - SensorLoader helper class to get sensor classes from files.   * Add - SensorContainerManager class to run sensors   * Fix - cmd/sensormanager.py now uses SensorLoader and SensorContainerManager     classes.
