+++ st2reactor/bin/sensor_container
@@ -18,4 +18,4 @@ if DEVEL:
     sys.path.append(python_lib_dir)
 
 if __name__ == '__main__':
-    cmd.sensorcontainer.main()
+    cmd.sensormanager.main()
