+++ st2reactor/st2reactor/cmd/__init__.py
@@ -1,3 +1,3 @@
-from st2reactor.cmd import sensorcontainer
+from st2reactor.cmd import sensormanager
 
-__all__ = ['sensorcontainer']
+__all__ = ['sensormanager']
